refactor: move `Parameter` to other ApiElements
